backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0
librosa==0.10.1
soundfile==0.12.1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.4
httpx>=0.25.0
praat-parselmouth>=0.4.3
spacy>=3.7.0
reportlab>=4.0.0
matplotlib>=3.8.0
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt==4.0.1
